ofmcc-4619 - add rate limit to backend
