+++ backend/package.json
@@ -39,6 +39,7 @@
     "dotenv": "^8.0.0",
     "express": "^4.18.2",
     "express-prometheus-middleware": "^0.8.5",
+    "express-rate-limit": "^7.4.0",
     "express-session": "^1.17.1",
     "express-validator": "^7.1.0",
     "fast-safe-stringify": "^2.0.7",
@@ -63,6 +64,7 @@
     "passport-openidconnect-keycloak-idp": "^0.1.6",
     "path": "0.12.7",
     "querystring": "0.2.0",
+    "rate-limit-redis": "^4.2.0",
     "redlock": "^4.2.0",
     "strip-ansi": "^6.0.0",
     "uuid": "^8.3.2",
