+++ backend/src/routes/public.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const { getSystemMessages } = require('../components/public')
+const { getSystemMessages } = require('../components/lookup')
 
 router.get('/system-messages', (req, res) => {
   getSystemMessages(req, res)
